Create Regression_Line
The dataset used in this algorithm contains the months of experience and salary of each employee in a given area. The linear regression algorithm was used to predict the salary of professionals in this area according to their length of experience in order to adapt the vacancy specifications to the salary range that the company intends to pay when hiring this employee.